bootstrap - move some methods that are only used once
